+++ CHANGELOG.md
@@ -6,8 +6,8 @@
 - none
 
 # v2023.8.1-beta
+- sqlmath - Add sql-function win_cosfit2_predict(), win_cosfit2_extract().
 - sqlmath - Add c-helper-functions doubleAbs(), doubleMax(), doubleMin().
-- sqlmath - Add sql-function win_cosfit2_predict().
 - sqlmath - Revert cosine-fit-update in sql-function win_cosfit2_step() due to poor performance in general-function win_cosfit2().
 - sqlmath - merge c-struct WinCosfitInternal, WinCosfitResult into c-struct WinCosfit.
 - sqlmath - Update sql-functions win_cosfit2(), win_emax(), win_quantilex() to move variable-length arguments to last position.